feat(): Enhance & make curve canvases consistent
