+++ src/common/utilities.ts
@@ -1,0 +1,4 @@
+/* [Exports] */
+export function degreesToRadians(degrees: number): number {
+  return (degrees / 360) * (2 * Math.PI);
+}
