+++ src/tabs/common/types.ts
@@ -1,0 +1,10 @@
+/* [Exports] */
+export type PlayButtonProps = {
+  isPlaying: boolean,
+  onClickCallback: () => void,
+};
+
+export type AutoLoopSwitchProps = {
+  isAutoLooping: boolean,
+  onChangeCallback: () => void,
+};
